GRHB-498: * moved sign out button to sidebar
