+++ mobile/src/common/types/ui/icon-name.type.ts
@@ -16,6 +16,7 @@ type IconName =
   | 'interview'
   | 'visibility'
   | 'visibilityOff'
+  | 'signOut'
   | 'send';
 
 export { type IconName };
