+++ mobile/src/navigation/app/components/drawer-content/drawer-content.tsx
@@ -8,6 +8,7 @@ import {
   RootScreenName,
 } from '~/common/enums/enums';
 import {
+  Button,
   Image,
   Link,
   SafeAreaView,
@@ -15,13 +16,13 @@ import {
   View,
 } from '~/components/common/common';
 import { getImageUri, groupByKey } from '~/helpers/helpers';
-import { useAppDispatch, useAppSelector } from '~/hooks/hooks';
+import { useAppDispatch, useAppNavigate, useAppSelector } from '~/hooks/hooks';
 import { DrawerNavigationItem } from '~/navigation/app/common/types/types';
 import {
   BecomeMentor,
   DrawerList,
 } from '~/navigation/app/components/components';
-import { coursesActions } from '~/store/actions';
+import { authActions, coursesActions } from '~/store/actions';
 
 import { styles } from './styles';
 
@@ -31,6 +32,7 @@ type Props = DrawerContentComponentProps & {
 
 const DrawerContent: FC<Props> = ({ state, items }) => {
   const dispatch = useAppDispatch();
+  const navigation = useAppNavigate();
   const focusedRouteName = state.routes[state.index].name as AppScreenName;
 
   const { user, isMentorBecomingVisible, dataBecomeMentorStatus } =
@@ -42,6 +44,13 @@ const DrawerContent: FC<Props> = ({ state, items }) => {
 
   const handleBecomeMentor = (): void => {
     dispatch(coursesActions.becomeMentor());
+  };
+
+  const handleLogout = async (): Promise<void> => {
+    await dispatch(authActions.signOut());
+    navigation.navigate(RootScreenName.AUTH, {
+      screen: AuthScreenName.SIGN_IN,
+    });
   };
 
   const itemGroups = groupByKey(items, 'drawerGroup');
@@ -62,6 +71,11 @@ const DrawerContent: FC<Props> = ({ state, items }) => {
             />
           </View>
         ))}
+        {user && (
+          <View style={styles.singOutWrapper}>
+            <Button label="Sign Out" icon="signOut" onPress={handleLogout} />
+          </View>
+        )}
         {isMentorBecomingVisible && user && (
           <BecomeMentor
             dataStatus={dataBecomeMentorStatus}
